requirements.txt
django==1.10.3
djangorestframework==3.5.3
coverage==4.3.1
factory_boy==2.8.1
django-debug-toolbar==1.6
django-debug-panel==0.8.3
